Move classes related to GNU dummy keys to gnupg package
