+++ pgpainless-core/src/main/java/org/gnupg/GnuPGDummyKeyUtil.java
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package org.pgpainless.key.gnu_dummy_s2k;
+package org.gnupg;
 
 import org.bouncycastle.bcpg.PublicKeyPacket;
 import org.bouncycastle.bcpg.S2K;
